@@ -1,6 +1,6 @@
 module github.com/ooni/oohttp
 
-go 1.20
+go 1.21
 
 require golang.org/x/net v0.22.0
 
