@@ -2,8 +2,7 @@
 // Use of this source code is governed by a BSD-style
 // license that can be found in the LICENSE file.
 
-//go:build !(js && wasm)
-// +build !js !wasm
+//go:build !wasm
 
 package http
 
